fix: pagination bug
If the user decides to change the issues filter (e.g. "all" to "open") and the current page (e.g. "all" 6) is higher than the next issues filter page (e.g. "open" max 5 pages) this will return an empty page for the user as if the research turns out null. This change forces a reset in pagination.